Add denylist for content emitters (#167)
Adds the ability to configure composables that probably emit content that we don't want to count towards the count for emissions. This is useful, for example, for composables that wrap stuff that will render in another window (like a Dialog or a Modal), which they do indeed emit content but we might not want to take it into account because it's not a mistake that is used at the same level as other stuff (e.g. a Scaffold composable).

Closes #166.